Move the SPI implementation outside of the impl package to follow the conventions
