+++ vertx-discovery-backend-redis/src/main/java/io/vertx/ext/discovery/backend/RedisBackend.java
@@ -14,7 +14,7 @@
  * You may elect to redistribute this code under either of these licenses.
  */
 
-package io.vertx.ext.discovery.backend.impl;
+package io.vertx.ext.discovery.backend;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Future;
