+++ vertx-discovery-backend-redis/src/main/resources/META-INF/services/io.vertx.ext.discovery.spi.DiscoveryBackend
@@ -14,4 +14,4 @@
 # You may elect to redistribute this code under either of these licenses.
 #
 
-io.vertx.ext.discovery.backend.impl.RedisBackend
+io.vertx.ext.discovery.backend.RedisBackend
